Added version 1.7.735

+++ chromium/components/update_client/component_unpacker.cc
@@ -25,6 +25,7 @@
 #include "components/update_client/component_patcher.h"
 #include "components/update_client/component_patcher_operation.h"
 #include "components/update_client/update_client.h"
+#include "components/update_client/update_client_errors.h"
 #include "crypto/secure_hash.h"
 #include "crypto/sha2.h"
 #include "third_party/zlib/google/zip.h"
@@ -33,24 +34,6 @@ using crypto::SecureHash;
 using crx_file::CrxFile;
 
 namespace update_client {
-
-ComponentUnpacker::ComponentUnpacker(
-    const std::vector<uint8_t>& pk_hash,
-    const base::FilePath& path,
-    const std::string& fingerprint,
-    const scoped_refptr<CrxInstaller>& installer,
-    const scoped_refptr<OutOfProcessPatcher>& oop_patcher,
-    const scoped_refptr<base::SequencedTaskRunner>& task_runner)
-    : pk_hash_(pk_hash),
-      path_(path),
-      is_delta_(false),
-      fingerprint_(fingerprint),
-      installer_(installer),
-      oop_patcher_(oop_patcher),
-      error_(kNone),
-      extended_error_(0),
-      task_runner_(task_runner) {
-}
 
 // TODO(cpu): add a specific attribute check to a component json that the
 // extension unpacker will reject, so that a component cannot be installed
@@ -72,6 +55,25 @@ std::unique_ptr<base::DictionaryValue> ReadManifest(
       static_cast<base::DictionaryValue*>(root.release()));
 }
 
+ComponentUnpacker::Result::Result() {}
+
+ComponentUnpacker::ComponentUnpacker(
+    const std::vector<uint8_t>& pk_hash,
+    const base::FilePath& path,
+    const scoped_refptr<CrxInstaller>& installer,
+    const scoped_refptr<OutOfProcessPatcher>& oop_patcher,
+    const scoped_refptr<base::SequencedTaskRunner>& task_runner)
+    : pk_hash_(pk_hash),
+      path_(path),
+      is_delta_(false),
+      installer_(installer),
+      oop_patcher_(oop_patcher),
+      error_(UnpackerError::kNone),
+      extended_error_(0),
+      task_runner_(task_runner) {}
+
+ComponentUnpacker::~ComponentUnpacker() {}
+
 bool ComponentUnpacker::UnpackInternal() {
   return Verify() && Unzip() && BeginPatching();
 }
@@ -79,13 +81,13 @@ bool ComponentUnpacker::UnpackInternal() {
 void ComponentUnpacker::Unpack(const Callback& callback) {
   callback_ = callback;
   if (!UnpackInternal())
-    Finish();
+    EndUnpacking();
 }
 
 bool ComponentUnpacker::Verify() {
   VLOG(1) << "Verifying component: " << path_.value();
   if (pk_hash_.empty() || path_.empty()) {
-    error_ = kInvalidParams;
+    error_ = UnpackerError::kInvalidParams;
     return false;
   }
   // First, validate the CRX header and signature. As of today
@@ -97,7 +99,7 @@ bool ComponentUnpacker::Verify() {
       path_, std::string(), &public_key_base64, nullptr, &header);
   if (error != CrxFile::ValidateError::NONE ||
       !base::Base64Decode(public_key_base64, &public_key_bytes)) {
-    error_ = kInvalidFile;
+    error_ = UnpackerError::kInvalidFile;
     return false;
   }
   is_delta_ = CrxFile::HeaderIsDelta(header);
@@ -112,7 +114,7 @@ bool ComponentUnpacker::Verify() {
 
   if (!std::equal(pk_hash_.begin(), pk_hash_.end(), hash)) {
     VLOG(1) << "Hash mismatch: " << path_.value();
-    error_ = kInvalidId;
+    error_ = UnpackerError::kInvalidId;
     return false;
   }
   VLOG(1) << "Verification successful: " << path_.value();
@@ -120,17 +122,18 @@ bool ComponentUnpacker::Verify() {
 }
 
 bool ComponentUnpacker::Unzip() {
+  // Mind the reference to non-const type, passed as an argument below.
   base::FilePath& destination = is_delta_ ? unpack_diff_path_ : unpack_path_;
-  VLOG(1) << "Unpacking in: " << destination.value();
   if (!base::CreateNewTempDirectory(base::FilePath::StringType(),
                                     &destination)) {
     VLOG(1) << "Unable to create temporary directory for unpacking.";
-    error_ = kUnzipPathError;
+    error_ = UnpackerError::kUnzipPathError;
     return false;
   }
+  VLOG(1) << "Unpacking in: " << destination.value();
   if (!zip::Unzip(path_, destination)) {
     VLOG(1) << "Unzipping failed.";
-    error_ = kUnzipFailed;
+    error_ = UnpackerError::kUnzipFailed;
     return false;
   }
   VLOG(1) << "Unpacked successfully";
@@ -142,7 +145,7 @@ bool ComponentUnpacker::BeginPatching() {
     // Use a different temp directory for the patch output files.
     if (!base::CreateNewTempDirectory(base::FilePath::StringType(),
                                       &unpack_path_)) {
-      error_ = kUnzipPathError;
+      error_ = UnpackerError::kUnzipPathError;
       return false;
     }
     patcher_ = new ComponentPatcher(unpack_diff_path_, unpack_path_, installer_,
@@ -153,63 +156,35 @@ bool ComponentUnpacker::BeginPatching() {
                    base::Bind(&ComponentUnpacker::EndPatching,
                               scoped_refptr<ComponentUnpacker>(this))));
   } else {
-    task_runner_->PostTask(
-        FROM_HERE,
-        base::Bind(&ComponentUnpacker::EndPatching,
-                   scoped_refptr<ComponentUnpacker>(this), kNone, 0));
+    task_runner_->PostTask(FROM_HERE,
+                           base::Bind(&ComponentUnpacker::EndPatching,
+                                      scoped_refptr<ComponentUnpacker>(this),
+                                      UnpackerError::kNone, 0));
   }
   return true;
 }
 
-void ComponentUnpacker::EndPatching(Error error, int extended_error) {
+void ComponentUnpacker::EndPatching(UnpackerError error, int extended_error) {
   error_ = error;
   extended_error_ = extended_error;
   patcher_ = NULL;
-  if (error_ != kNone) {
-    Finish();
-    return;
-  }
-  // Optimization: clean up patch files early, in case disk space is too low to
-  // install otherwise.
-  if (!unpack_diff_path_.empty()) {
-    base::DeleteFile(unpack_diff_path_, true);
-    unpack_diff_path_.clear();
-  }
-  Install();
-  Finish();
+
+  EndUnpacking();
 }
 
-void ComponentUnpacker::Install() {
-  // Write the fingerprint to disk.
-  if (static_cast<int>(fingerprint_.size()) !=
-      base::WriteFile(
-          unpack_path_.Append(FILE_PATH_LITERAL("manifest.fingerprint")),
-          fingerprint_.c_str(), base::checked_cast<int>(fingerprint_.size()))) {
-    error_ = kFingerprintWriteFailed;
-    return;
-  }
-  std::unique_ptr<base::DictionaryValue> manifest(ReadManifest(unpack_path_));
-  if (!manifest.get()) {
-    error_ = kBadManifest;
-    return;
-  }
-  DCHECK(error_ == kNone);
-  if (!installer_->Install(*manifest, unpack_path_)) {
-    error_ = kInstallerError;
-    return;
-  }
-}
-
-void ComponentUnpacker::Finish() {
+void ComponentUnpacker::EndUnpacking() {
   if (!unpack_diff_path_.empty())
     base::DeleteFile(unpack_diff_path_, true);
-  if (!unpack_path_.empty())
+  if (error_ != UnpackerError::kNone && !unpack_path_.empty())
     base::DeleteFile(unpack_path_, true);
-  task_runner_->PostTask(FROM_HERE,
-                         base::Bind(callback_, error_, extended_error_));
-}
 
-ComponentUnpacker::~ComponentUnpacker() {
+  Result result;
+  result.error = error_;
+  result.extended_error = extended_error_;
+  if (error_ == UnpackerError::kNone)
+    result.unpack_path = unpack_path_;
+
+  task_runner_->PostTask(FROM_HERE, base::Bind(callback_, result));
 }
 
 }  // namespace update_client
